# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v0/companies": {
      "post": {
        "operationId": "CompaniesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (details lists the field messages)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Logto token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Trial expired (only on routes not allowed after expiry)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Company not registered yet, or resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Usuário já tem empresa ou CNPJ já cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "logto": []
          }
        ],
        "summary": "Cria a empresa do usuário (uma vez); inicia o trial de 30 dias",
        "tags": [
          "companies"
        ]
      }
    },
    "/api/v0/companies/me": {
      "get": {
        "operationId": "CompaniesController_findMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (details lists the field messages)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Logto token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Trial expired (only on routes not allowed after expiry)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Company not registered yet, or resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "logto": []
          }
        ],
        "summary": "Empresa do usuário (404 = ainda não cadastrou → onboarding)",
        "tags": [
          "companies"
        ]
      },
      "patch": {
        "operationId": "CompaniesController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (details lists the field messages)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Logto token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Trial expired (only on routes not allowed after expiry)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Company not registered yet, or resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "logto": []
          }
        ],
        "summary": "Atualiza dados da empresa (cnpj não muda)",
        "tags": [
          "companies"
        ]
      }
    },
    "/api/v0/companies/me/certificate": {
      "put": {
        "operationId": "CompaniesController_setCertificate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file",
                  "password"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Certificado A1 (.pfx, máx. 10 KB)"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (details lists the field messages)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Logto token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Trial expired (only on routes not allowed after expiry)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Company not registered yet, or resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Arquivo inválido ou senha errada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "logto": []
          }
        ],
        "summary": "Envia o certificado A1 (.pfx + senha)",
        "tags": [
          "companies"
        ]
      }
    },
...
  },
  "info": {
    "title": "MATHEO NFS-e API",
    "description": "Emissor de NFS-e para MEI. Todas as rotas exigem Bearer token do Logto (audience https://api.matheo.local), exceto /health e esta documentação. Erros seguem { statusCode, message, details?, requestId }. Ver docs/frontend/api-brief.md para fluxos e estados.",
    "version": "0.3.2",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "logto": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 422
          },
          "message": {
            "type": "string",
            "example": "NFS-e rejected by the national API"
          },
          "details": {
            "description": "Validation: array of field messages. Business errors: object with context (e.g. { invoiceId, code, reason }).",
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              {
                "type": "object",
                "additionalProperties": true
              }
            ]
          },
          "requestId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-4a7b-8c9d-0e1f2a3b4c5d",
            "description": "Same value as the x-request-id response header"
          }
        },
        "required": [
          "statusCode",
          "message",
          "requestId"
        ]
      },
      "CreateCompanyDto": {
        "type": "object",
        "properties": {
          "cnpj": {
            "type": "string",
            "example": "11222333000181",
            "description": "CNPJ da empresa, 14 dígitos, com dígito verificador válido"
          },
          "razaoSocial": {
            "type": "string",
            "example": "Minha MEI LTDA",
            "minLength": 2,
            "maxLength": 150
          },
          "inscricaoMunicipal": {
            "type": "string",
            "example": "123456",
            "maxLength": 20
          },
          "codigoMunicipio": {
            "type": "string",
            "example": "3550308",
            "description": "Código IBGE do município (7 dígitos)"
          },
          "email": {
            "type": "string",
            "example": "contato@exemplo.com"
          },
          "telefone": {
            "type": "string",
            "example": "11999999999",
            "description": "10 ou 11 dígitos"
          }
        },
        "required": [
          "cnpj",
          "razaoSocial",
          "codigoMunicipio"
        ]
      },
      "CompanyResponseModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "cnpj": {
            "type": "string",
            "example": "11222333000181"
          },
          "razaoSocial": {
            "type": "string",
            "example": "Minha MEI LTDA"
          },
          "inscricaoMunicipal": {
            "type": "string",
            "nullable": true,
            "example": "123456"
          },
          "codigoMunicipio": {
            "type": "string",
            "example": "3550308"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "example": "contato@exemplo.com"
          },
          "telefone": {
            "type": "string",
            "nullable": true,
            "example": "11999999999"
          },
          "hasCertificate": {
            "type": "boolean",
            "description": "Whether an A1 certificate is stored (the certificate itself is never returned)"
          },
          "certificateExpiry": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "plan": {
            "type": "string",
            "enum": [
              "TRIAL",
              "ACTIVE"
            ]
          },
          "trialEndsAt": {
            "type": "string",
            "format": "date-time",
            "description": "End of the 30-day trial, counted from POST /companies"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "cnpj",
          "razaoSocial",
          "inscricaoMunicipal",
          "codigoMunicipio",
          "email",
          "telefone",
          "hasCertificate",
          "certificateExpiry",
          "plan",
          "trialEndsAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateCompanyDto": {
        "type": "object",
        "properties": {
          "razaoSocial": {
            "type": "string",
            "example": "Minha MEI LTDA",
            "minLength": 2,
            "maxLength": 150
          },
          "inscricaoMunicipal": {
            "type": "string",
            "example": "123456",
            "maxLength": 20
          },
          "codigoMunicipio": {
            "type": "string",
            "example": "3550308",
            "description": "Código IBGE do município (7 dígitos)"
          },
          "email": {
            "type": "string",
            "example": "contato@exemplo.com"
          },
          "telefone": {
            "type": "string",
            "example": "11999999999",
            "description": "10 ou 11 dígitos"
          }
        }
      },
...
    }
  }
}